renamed and reorganized files to better illustrate messaging system
