+++ src/Rocket3d/input/messaging/send/AppInputSender.cpp
@@ -1,0 +1,65 @@
+#include "AppInputSender.h"
+#include "../../../../Rocket3d/core/Application.h"
+#include "../../../../Rocket3d/input/messaging/Message.h"
+#include "../receive/AppMessages.h"
+#include <glfw3.h>
+
+AppInputSender::AppInputSender()
+{
+}
+
+AppInputSender::~AppInputSender()
+{
+}
+
+bool AppInputSender::handleKeyEvents(int key, int scancode, int action, int mods)
+{
+	bool shouldConsumeEvent = false;
+
+	onKeyDown(key, scancode, action, mods);
+
+	return shouldConsumeEvent;
+}
+
+bool AppInputSender::handleMouseButtonEvents(int key, int scancode, int action, int mods)
+{
+	bool shouldConsumeEvent = false;
+
+	return shouldConsumeEvent;
+}
+
+void AppInputSender::onKeyDown(int key, int scancode, int action, int mods)
+{
+	if (key == GLFW_KEY_ENTER && action == GLFW_PRESS)
+	{
+		Message* pMessage = new AppKeyDown(ENTER);
+		Application::getInstance()->getMessageManager()->addMessage(pMessage, 1);
+	}
+
+	if (key == GLFW_KEY_ESCAPE && action == GLFW_PRESS)
+	{
+		Message* pMessage = new AppKeyDown(ESC);
+		Application::getInstance()->getMessageManager()->addMessage(pMessage, 1);
+	}
+}
+
+void AppInputSender::onKeyUp(int key, int scancode, int action, int mods)
+{
+}
+
+void AppInputSender::onKeyRepeat(int key, int scancode, int action, int mods)
+{
+}
+
+void AppInputSender::onMouseDown(int key, int scancode, int action, int mods)
+{
+}
+
+void AppInputSender::onMouseUp(int key, int scancode, int action, int mods)
+{
+}
+
+void AppInputSender::onMouseRepeat(int key, int scancode, int action, int mods)
+{
+}
+
