+++ src/RocketEngine/RocketEngine.vcxproj.filters
@@ -72,7 +72,7 @@
     <ClInclude Include="platform\OpenGL\ImGuiOpenGLRenderer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="input\InputInterface.h">
+    <ClInclude Include="input\InputSender.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
